More big infrastructure changes. Beginning of a testing suite that tests several different arguments to the and special form.
