--- evaluation/eval_test.go
@@ -1,10 +1,0 @@
-package evaluation
-
-import (
-	"testing"
-)
-
-func TestAnd(t *testing.T) {
-	test1 := and(t)
-
-}
